Refactor IKSolver.cpp into separate solver files
Split the monolithic IKSolver.cpp (1828 lines) into focused files:
- IKUtils.cpp: Transform utilities (compose/decompose, aimAt)
- TwoBoneIKSolver.cpp: Two-bone IK solver for arms/legs
- LookAtIKSolver.cpp: Look-at IK for head/eyes
- FootPlacementIKSolver.cpp: Foot placement with ground queries
- StraddleIKSolver.cpp: Hip tilt/shift for uneven terrain
- ClimbingIKSolver.cpp: Climbing IK with hand/foot holds

IKSolver.cpp now only contains IKSystem orchestration (~420 lines).
